fix implements_design_check carrier
