+++ src/faebryk/library/is_pulled.py
@@ -1,19 +1,7 @@
 import faebryk.library._F as F
 from faebryk.core.module import Module
-from faebryk.core.node import Node
 from faebryk.core.pathfinder import find_paths
-from faebryk.library.implements_design_check import CheckException
-from faebryk.libs.library import L
 from faebryk.libs.util import once
-
-
-class RequiresPullNotFulfilled(CheckException):
-    def __init__(self, nodes: list[Node]):
-        self.nodes = nodes
-        super().__init__(
-            f"Signals requiring pulls but not pulled: "
-            f"{', '.join(mif.get_full_name() for mif in nodes)}"
-        )
 
 
 class is_pulled(Module.TraitT.decless()):
@@ -21,17 +9,13 @@ class is_pulled(Module.TraitT.decless()):
         super().__init__()
         self.signal = signal
 
-    @L.rt_field
-    def check(self) -> F.implements_design_check:
-        @once  # TODO: only cache once True
-        def _check():
-            # TODO: more efficient method?
-            paths = find_paths(self.signal, [self.signal.reference])
-            for path in paths:
-                for node in path:
-                    if isinstance(node, F.Resistor):
-                        return
+    @once  # TODO: only cache once True
+    def check(self) -> bool:
+        # TODO: more efficient method?
+        paths = find_paths(self.signal, [self.signal.reference])
+        for path in paths:
+            for node in path:
+                if isinstance(node, F.Resistor):
+                    return True
 
-            raise RequiresPullNotFulfilled([self.signal])
-
-        return F.implements_design_check(_check)
+        return False
